requirements.txt
# Aiogram - Modern async Telegram bot framework
aiogram

# Async HTTP client
aiohttp

# Async PostgreSQL driver
asyncpg

# ORM for database
SQLAlchemy

# Scheduling
APScheduler

# Configuration
python-dotenv

# HTTP requests (for prayer API)
requests